### 2024-02-29 6708f1e
add missing bioconda doi
