### 2024-05-15 #6333 @JeanMainguy
Hi @luizirber

Binette submission seems to be stuck in pre-review. Is there anything I can do to move it forward?
In the meantime, I've uploaded Binette to BioRxiv (https://www.biorxiv.org/content/10.1101/2024.04.20.585171v1).

Thanks!
