### 2024-05-19 f410f95
improve error management when bin contigs are missing in contig file  error from issue #11
