### 2024-05-23 #6333 @arfon
@lskatz, @oschwengers, @beardymcjohnface, @Robaina, @mdehollander – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is Binette: a fast and accurate bin refinement tool to construct high quality Metagenome Assembled Genomes.

Based on your experience, and past submissions to JOSS, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
