### 2024-05-25 #6782 @editorialbot
[REVIEW]: Binette: a fast and accurate bin refinement tool to construct high quality Metagenome Assembled Genomes.
**Submitting author:** @JeanMainguy (<a href="http://orcid.org/0009-0006-9160-9744">Jean Mainguy</a>)
**Repository:** https://github.com/genotoul-bioinfo/Binette
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
