### 2024-06-13 genotoul-bioinfo/Binette#15 @lskatz
Unit tests
Hi, could you help me understand how to run the unit tests locally? This is part of the review at <https://github.com/openjournals/joss-reviews/issues/6782>.
