### 2024-06-14 genotoul-bioinfo/Binette#15 @JeanMainguy
Hi @Iskatz,

Absolutely, happy to help!

### Set Up Your Environment

You’ll need to get Binette’s dependencies installed. The easiest way is with conda or mamba. Just run:

```bash
mamba create -c bioconda -c defaults -c conda-forge -n binette_test binette
conda activate binette_test
```

### Clone the Repository

Next, grab the Binette repo and navigate into it:

```bash
git clone https://github.com/genotoul-bioinfo/Binette
cd Binette
```

### Install Binette and Dev Dependencies

Now, install Binette along with its dev dependencies (like `pytest` and `pytest-cov`) using pip:

```bash
pip install .[dev]
```

### Run the Unit Tests

Finally, you can run the unit tests with:

```bash
pytest .
```
Feel free to reach out if you hit any problem!
