### 2024-06-14 genotoul-bioinfo/Binette#15 @JeanMainguy
I just tried the commands above, and I ran into an issue with `pyfastx`. Conda downloads a very old version that doesn't work correctly. I'll need to fix the conda recipe to force the correct `pyfastx` version.

If you encounter the same problem, you can simply update `pyfastx` with conda or mamba:

```bash
mamba install 'pyfastx>=2'
```

or with pip directly:

```bash
pip install "pyfastx>=2"
```
