### 2024-06-14 genotoul-bioinfo/Binette#15 @JeanMainguy
The conda recipe is now fixed. I tried the commands and its working as expected 👌
