### 2024-06-16 ffad8f4
add extension checks to only look for fasta files in bin dirs input
