### 2024-06-16 ca1df58
test bin dirs input in CI
