+++ .github/workflows/binette_ci.yml
@@ -72,13 +72,26 @@ jobs:
         repository: genotoul-bioinfo/Binette_TestData
         path: test_data
 
-    - name: Run simple test case
+    - name: Run simple test case from bin2table
       run: |
         cd test_data
         binette -b binning_results/* --contigs all_contigs.fna --checkm2_db checkm2_tiny_db/checkm2_tiny_db.dmnd  -v -o test_results
 
-    - name: Compare results with expectation
+    - name: Compare results from bin2table with expectation
       run: |
         cd test_data
         head  expected_results/final_bins_quality_reports.tsv test_results/final_bins_quality_reports.tsv
         python scripts/compare_results.py expected_results/final_bins_quality_reports.tsv test_results/final_bins_quality_reports.tsv
+
+
+    - name: Run simple test case from bin dirs
+      run: |
+        cd test_data
+        binette -d binning_results/A/ binning_results/B/ binning_results/C/  --contigs all_contigs.fna --checkm2_db checkm2_tiny_db/checkm2_tiny_db.dmnd  -v -o test_results_from_dirs
+
+    - name: Compare results from bin dirs with expectation
+      run: |
+        cd test_data
+        head  expected_results/final_bins_quality_reports.tsv test_results_from_dirs/final_bins_quality_reports.tsv
+        python scripts/compare_results.py expected_results/final_bins_quality_reports.tsv test_results_from_dirs/final_bins_quality_reports.tsv
+
