### 2024-06-16 2e5418d
fix cmds in CI
