### 2024-06-21 genotoul-bioinfo/Binette#18 @lskatz
Sample dataset
Hi, this is part of the review for https://github.com/openjournals/joss-reviews/issues/6782

I was trying to follow the examples manually but I do not see any example data. I got confused by trying to use your examples and then getting errors. But I think those errors have to do with a mismatch between bins and what is actually in the fasta file. Could you add example data to the repo with examples on how to use it? And then what we'd expect as output for each specific example.

In more detail, I believe that we have a mismatch while following the current example and so I think you should put the proper example data into the repo with instructions.

```text
(binette) $ mkdir LKTEST
(binette) $ cd LKTEST/
(binette) $ cat > bin_set1.tsv
contig_1   binA
contig_8   binA
contig_15  binB
contig_9   binC
(binette) $ cat > bin_set2.tsv
contig_1   bin.0
contig_8   bin.0
contig_15  bin.1
contig_9   bin.2
contig_10  bin.0
(binette) $ find .. -name '*.fasta'
../tests/contigs.fasta
../output.fasta
(binette) $ binette --contig2bin_tables bin_set1.tsv bin_set2.tsv --contigs ../tests/contigs.fasta
Traceback (most recent call last):
  File "/bin/miniconda3/envs/binette/bin/binette", line 8, in <module>
    sys.exit(main())
  File "/bin/miniconda3/envs/binette/lib/python3.8/site-packages/binette/main.py", line 334, in main
    bin_set_name_to_bins, original_bins, contigs_in_bins, contig_to_length = parse_input_files(args.bin_dirs, args.contig2bin_tables, args.contigs, fasta_extensions=set(args.fasta_extensions))
  File "/bin/miniconda3/envs/binette/lib/python3.8/site-packages/binette/main.py", line 153, in parse_input_files
    bin_set_name_to_bins = bin_manager.parse_contig2bin_tables(bin_name_to_bin_table)
  File "/bin/miniconda3/envs/binette/lib/python3.8/site-packages/binette/bin_manager.py", line 222, in parse_contig2bin_tables
    bin_name_to_bins[name] = get_bins_from_contig2bin_table(contig2bin_table, name)
  File "/bin/miniconda3/envs/binette/lib/python3.8/site-packages/binette/bin_manager.py", line 243, in get_bins_from_contig2bin_table
    bin_name = line.strip().split("\t")[1]
IndexError: list index out of range
(binette) $ seqtk comp ../tests/contigs.fasta
contig1 1020    294     190     286     250     0       0       0       110     0       0       0
contig2 1680    598     248     292     542     0       0       0       26      0       0       0

```
