### 2024-06-22 genotoul-bioinfo/Binette#18 @lskatz
Ok I am going to try some [real data](https://github.com/ncezid-biome/datasets?tab=readme-ov-file#edlb) from our metagenomics benchmark. Do you have suggestions on how to make good input data MAGs for Binette? I was thinking of trying out [SemiBin2](https://github.com/BigDataBiology/SemiBin_tutorial_from_scratch).
