### 2024-06-24 genotoul-bioinfo/Binette#18 @lskatz
Hi, I tried to run Binette in full but I don't really know what's wrong. Can you look over this?

Steps:

I downloaded the metagenome `Kickstart` from the above dataset (SAMN05024035).

I ran megahit to get an assembly

```
megahit -1 Kickstart_1.fastq -2 Kickstart_2.fastq --out-dir Kickstart.megahit --out-prefix R1 --num-cpu-threads 12
```

I mapped the reads for input for SemiBin2 to create `Kickstart.megahit.sorted.bam`.

```
ref=Kickstart.megahit/R1.contigs.fa
R1=Kickstart_1.fastq
R2=${R1/_1/_2}
bowtie2-build -f $ref $ref
bowtie2 -q --fr -x $ref -1 $R1 -2 $R2 -S Kickstart.sam -p 12
samtools view -h -F 4 -b -S -o Kickstart.unsorted.bam -@ 12 Kickstart.sam
samtools sort -m 1000000000 Kickstart.unsorted.bam -o Kickstart.megahit.sorted.bam -@ 12

samtools flagstats Kickstart.megahit.sorted.bam
6275427 + 0 in total (QC-passed reads + QC-failed reads)
6275427 + 0 primary
0 + 0 secondary
0 + 0 supplementary
0 + 0 duplicates
0 + 0 primary duplicates
6275427 + 0 mapped (100.00% : N/A)
6275427 + 0 primary mapped (100.00% : N/A)
6275427 + 0 paired in sequencing
3160305 + 0 read1
3115122 + 0 read2
5859266 + 0 properly paired (93.37% : N/A)
6146726 + 0 with itself and mate mapped
128701 + 0 singletons (2.05% : N/A)
49282 + 0 with mate mapped to a different chr
34450 + 0 with mate mapped to a different chr (mapQ>=5)

```

I ran SemiBin2 three times to give me different results/inputs for Binette. Each iteration uses a random seed and gives a different output.

```
for rep in 1 2 3; do
  SemiBin2 single_easy_bin -i $ref -b Kickstart.megahit.sorted.bam -o Kickstart.megahit.semibin$rep -p 12 > Kickstart.megahit.semibin.log 2>&1;
done
```

For Binette, I prepared an input folder and then ran it

```
for i in *megahit.semibin*; do
  # Remove header
  tail -n +2 $i/contig_bins.tsv > binette.tmp/$i.tsv;
done;
# oops it tried to read from the log file
rm binette.tmp/Kickstart.megahit.semibin.log.tsv
ls binette.tmp/ -lh
total 242K
-rw-------. 1 user users 81K Jun 24 10:50 Kickstart.megahit.semibin2.tsv
-rw-------. 1 user users 81K Jun 24 10:50 Kickstart.megahit.semibin3.tsv
-rw-------. 1 user users 81K Jun 24 10:50 Kickstart.megahit.semibin.tsv
$ head binette.tmp/*
==> binette.tmp/Kickstart.megahit.semibin2.tsv <==
k141_85796      0
k141_53630      1
k141_21463      2
k141_85804      3
k141_32191      4
k141_42940      5
k141_75081      6
k141_10736      7
k141_42942      3
k141_32194      1

==> binette.tmp/Kickstart.megahit.semibin3.tsv <==
k141_85796      0
k141_53630      1
k141_21463      2
k141_85804      1
k141_32191      3
k141_42940      4
k141_75081      5
k141_10736      6
k141_42942      1
k141_32194      3

==> binette.tmp/Kickstart.megahit.semibin.tsv <==
k141_85796      0
k141_53630      1
k141_21463      2
k141_85804      1
k141_32191      3
k141_42940      4
k141_75081      5
k141_10736      6
k141_42942      7
k141_32194      1
```

So at this point it appears I have different results of MAGs for input for Binette, and I'm ready to run it.

```
binette --contig2bin_tables binette.tmp/Kickstart.megahit.semibin.tsv --contig2bin_tables binette.tmp/Kickstart.megahit.semibin2.tsv --contig2bin_tables binette.tmp/Kickstart.megahit.semibin3.tsv  --threads 12 --outdir Kickstart.megahit.semibin.1.2.3.binette --contigs $ref --checkm2_db $HOME/db/checkm2/CheckM2_database/uniref100.KO.1.dmnd >& binette.log &

tail binette.log
100%|██████████| 5945/5945 [00:00<00:00, 13598.82it/s]
100%|██████████| 5945/5945 [00:00<00:00, 244014.77contig/s]
100%|██████████| 5945/5945 [00:00<00:00, 574979.53contig/s]

head Kickstart.megahit.semibin.1.2.3.binette/final_bins_quality_reports.tsv
bin_id  origin  name    completeness    contamination   score   size    N50     contig_count
wc -l Kickstart.megahit.semibin.1.2.3.binette/final_bins_quality_reports.tsv
1 Kickstart.megahit.semibin.1.2.3.binette/final_bins_quality_reports.tsv

```

I think that the results are empty and so what should I do next?
