### 2024-06-24 genotoul-bioinfo/Binette#18 @JeanMainguy
Hi,

Thanks for testing the tool with real data!

It looks like the issue is with the `--contig2bin_tables` argument. You should specify it once and list all the files after it.

The correct command should be:

```bash
binette --contig2bin_tables binette.tmp/Kickstart.megahit.semibin.tsv binette.tmp/Kickstart.megahit.semibin2.tsv binette.tmp/Kickstart.megahit.semibin3.tsv \
        --threads 12 \
        --outdir Kickstart.megahit.semibin.1.2.3.binette \
        --contigs $ref \
        --checkm2_db $HOME/db/checkm2/CheckM2_database/uniref100.KO.1.dmnd >& binette.log &
```

If you check the beginning of the log file, it details the different files it processes:

So for example with the test data set:
```
[..]
[2024-06-24 19:30:40] INFO - Parsing bin2contig files.
[2024-06-24 19:30:40] INFO - 3 bin sets processed:
[2024-06-24 19:30:40] INFO -  A - 6 bins
[2024-06-24 19:30:40] INFO -  B - 3 bins
[2024-06-24 19:30:40] INFO -  C - 4 bins
```

In your case, only the last table is being processed.

This is something I surely should improve; the tool should ideally throw an error when the same argument is given multiple times.
