### 2024-06-24 genotoul-bioinfo/Binette#18 @lskatz
Thank you! This worked! Here is the result

```
bin_id  origin    name           completeness  contamination  score              size     N50    contig_count
23                22             100.0         0.09           99.82              4698156  82084  97
15256   union     18 | 17        99.93         0.28           99.37              2940499  37523  114
16043   union     25 | 27 | 144  96.2          0.47           95.26              4284705  39217  144
15592   union     21 | 25        93.93         0.19           93.55000000000001  2190965  12519  222
15577   union     34 | 104 | 12  85.47         1.49           82.49              2126765  10106  283
4076    diff      9 - 27         84.57         1.99           80.58999999999999  2971038  6934   492
1253    intersec  5 & 5          82.26         1.62           79.02000000000001  1643026  8518   241
16531   union     56 | 17        75.11         2.73           69.65              1434469  5017   294
14118   union     4 | 3 | 1      63.6          3.84           55.92              3743529  4716   805
15866   union     33 | 10        57.64         2.4            52.84              1892146  4848   410
4173    diff      34 - 14 - 3    48.26         0.49           47.28              1936962  4159   470
7531    diff      18 - 16        44.89         0.14           44.61              965464   4748   215
```
