### 2024-06-24 genotoul-bioinfo/Binette#18 @lskatz
Why does bin `23` transform into bin `22` in the first line?

Is there a suggested score threshold?
