### 2024-06-24 #6782 @lskatz
Some general comments on the paper

* Line 42: why is the weight set as 2?
* Clarity on line 45: currently is " rather than CheckM1 as in the metaWRAP pipeline" but maybe " rather than CheckM1, which is what the metaWRAP pipeline uses"
* Line 96: does JOSS allow for in preparation citations? (might be a question for the editor)
