### 2024-06-24 genotoul-bioinfo/Binette#19 @lskatz
Better tutorial/data
Hi, this is coming off of #18 and as a comment on the review for JOSS <https://github.com/openjournals/joss-reviews/issues/6782>.

* Please include real data in the repo in one way or another
* Please include a tutorial on how to use real data so that a user can follow it from beginning to end. One real datum that you could use is the one in #18 or maybe something from the CAMI benchmark datasets.
* Hook the real data into GitHub Actions even if it's in a minor way
