### 2024-06-25 genotoul-bioinfo/Binette#18 @JeanMainguy
The "bin id" column represents a unique identifier assigned by Binette, while the "name" column shows the original names of the bins that contributed to the final bin.
Here, the original name is "22" and the Binette assigned name is "23". This bin has not been combined with any other bins so the name is simply the original name.
The "name" column may not be very important here; it simply provides information on the initial bins that contributed to its creation and the way they contributed.
For instance `18 | 17 ` means the final bin has been made from the union of bin 18 and 17.
I understand that this might seem confusing and the name actually give you no clue from which bin set they are from especially if all bins are name the same way in all bin sets.
