### 2024-06-25 genotoul-bioinfo/Binette#19 @JeanMainguy
Hi,

I agree that would be important to provide a clear tutorial on how to use Binette with real data. I'll look into the best ways to include this in the repository. I'll probably add a new section in the documentation.

However, for GitHub Actions, I think we should stick to basic examples that demonstrate the full process without the size of real data. The main goal of the GitHub CI pipeline is to run tests when code changes, ensuring everything works as expected. Since these tests might run often, it's best to keep them quick to minimize computing time. So, I think the current [test dataset](https://github.com/genotoul-bioinfo/Binette_TestData) fulfills this purpose well.
