### 2024-06-25 genotoul-bioinfo/Binette#19 @lskatz
Understood! I will strike the suggestion for using real data in github actions, but I would appreciate a clear tutorial using real data.
