### 2024-06-25 genotoul-bioinfo/Binette#20 @lskatz
Documentation on seeking help
For the JOSS review at <https://github.com/openjournals/joss-reviews/issues/6782>.

According to the guidelines, can you add in a statement on how to seek help or support for the software? Thank you!

![image](https://github.com/genotoul-bioinfo/Binette/assets/14798/1cb09e49-ad1a-4d71-a3a9-7ee6d1539413)
