### 2024-06-28 43e30ed
improve doc to clarify step when seeking for help
