### 2024-07-01 genotoul-bioinfo/Binette#20 @JeanMainguy
Hi,

Thanks for the feedback!
Support is provided through the GitHub issues page. I've added info to clarify that to both the README ([here](https://github.com/genotoul-bioinfo/Binette?tab=readme-ov-file#help-feature-requests-and-bug-reporting)) and the documentation ([here](https://binette.readthedocs.io/en/latest/contributing.html#help-or-reporting-a-bug)).

Best,
