### 2024-07-07 #6782 @lskatz
> Any questions/concerns please let @arfon know.

@arfon does JOSS allow for in preparation citations?
