### 2024-07-07 genotoul-bioinfo/Binette#20 @lskatz
Perfect. There are a couple of questions left here <https://github.com/openjournals/joss-reviews/issues/6782#issuecomment-2187193217> and then a tutorial via #19 and then I think I can accept!
