### 2024-07-08 #6782 @arfon
> @arfon does JOSS allow for in preparation citations?

I think that's fine.
