### 2024-07-10 2514457
Run CI on macOS runner as well
