+++ .github/workflows/binette_ci.yml
@@ -13,7 +13,7 @@ on:
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    os: ['ubuntu-latest', 'macos-13']
     defaults:
       run:
         shell: bash -el {0}
