### 2024-07-10 genotoul-bioinfo/Binette#22 @JeanMainguy
Hi,

Unfortunately, I'm not able to reproduce the error on my end.

I ran the CI on a macOS runner via GitHub to see if the problem might be OS-related. The CI passes without any issues (check it out here: [CI Run](https://github.com/genotoul-bioinfo/Binette/actions/runs/9875187884/job/27271263003)). However, the GitHub macOS version is 13, which is lower than your version.

The error in your log seems to come from this line:
```
/bin/sh: line 1: 20164 Illegal instruction: 4  diamond blastp --outfmt 6 --max-target-seqs 1 --query test_results/temporary_files/assembly_proteins.faa -o test_results/temporary_files/diamond_result.tsv --threads 1 --db checkm2_tiny_db/checkm2_tiny_db.dmnd --query-cover 80 --subject-cover 80 --id 30 --evalue 1e-05 --block-size 2 2> test_results/temporary_files/diamond_result.log
```

I found this Stack Overflow question: [What is the illegal instruction 4 error and why does it happen?](https://stackoverflow.com/questions/14268887/what-is-the-illegal-instruction-4-error-and-why-does-mmacosx-version-min-10) and this GitHub issue: [DIAMOND Issue #303](https://github.com/bbuchfink/diamond/issues/303).

It seems the issue might be related to the DIAMOND conda package not being compatible with your version of macOS.
