### 2024-07-18 #6782 @Kevin-Mattheus-Moerman
@lskatz I would only add that if you can post a pre-print for these and cite these that would be preferred.
