### 2024-07-19 genotoul-bioinfo/Binette#22 @beardymcjohnface
Hi Jean,
That makes sense. I have not tried the instructions for installing a newer version of diamond, but Binette works perfectly on my linux system!
