### 2024-07-22 7fe9dd7
improve clarity as suggested in the JOSS review
see : https://github.com/openjournals/joss-reviews/issues/6782#issuecomment-2187193217
