### 2024-07-22 #6782 @JeanMainguy
Hello,
Thanks for the feedbacks.

>     * Line 42: why is the weight set as 2?

The contamination weight is set to 2 by default based on tests conducted with the CAMI dataset. I experimented with various weights and selected the optimal one. From my observations, this parameter does have a huge impact on the number of high-quality bins. The weight is adjustable and depends on the user's tolerance for contamination in the resulting bins.  It can be increased if the user is uncomfortable with some contamination in the final bins and wants to minimize it as much as possible.

>     * Clarity on line 45: currently is " rather than CheckM1 as in the metaWRAP pipeline" but maybe " rather than CheckM1, which is what the metaWRAP pipeline uses"

I've made the change. Thanks.

>     * Line 96: does JOSS allow for in preparation citations? (might be a question for the editor)

We are almost ready to submit the metagWGS paper. We will post it as a preprint and cite it accordingly as suggested by @Kevin-Mattheus-Moerman.
