### 2024-07-22 #6782 @JeanMainguy
Ok, great !
I'm working on the tutorial, but I'm having a bit of trouble figuring out the right dataset size. I want it to run on a laptop without taking forever, but still give some interesting results. But any way, I'll try to finish it soon !
