### 2024-08-28 55f7f55
add some warning if contig are duplicated in an input set and use id in bin name
