+++ binette/bin_manager.py
@@ -1,5 +1,4 @@
 import logging
-import os
 from collections import defaultdict
 from pathlib import Path
 
@@ -22,7 +21,7 @@ class Bin:
         """
         Bin.counter += 1
 
-        self.origin = origin
+        self.origin = {origin}
         self.name = name
         self.id = Bin.counter
         self.contigs = set(contigs)
@@ -37,7 +36,7 @@ class Bin:
 
         self.is_original = is_original
 
-    def __eq__(self, other: 'Bin') -> bool:
+    def __eq__(self, other) -> bool:
         """
         Compare the Bin object with another object for equality.
 
@@ -60,7 +59,7 @@ class Bin:
 
         :return: The string representation of the Bin object.
         """
-        return f"{self.origin}_{self.id}  ({len(self.contigs)} contigs)"
+        return f"Bin {self.id} from {';'.join(self.origin)}  ({len(self.contigs)} contigs)"
 
     def overlaps_with(self, other: 'Bin') -> Set[str]:
         """
@@ -71,18 +70,18 @@ class Bin:
         """
         return self.contigs & other.contigs
 
-    def __and__(self, other: 'Bin') -> 'Bin':
-        """
-        Perform a logical AND operation between this bin and another bin.
+    # def __and__(self, other: 'Bin') -> 'Bin':
+    #     """
+    #     Perform a logical AND operation between this bin and another bin.
 
-        :param other: The other Bin object.
-        :return: A new Bin object representing the intersection of the bins.
-        """
-        contigs = self.contigs & other.contigs
-        name = f"{self.name} & {other.name}"
-        origin = f"{self.origin} & {other.origin}"
+    #     :param other: The other Bin object.
+    #     :return: A new Bin object representing the intersection of the bins.
+    #     """
+    #     contigs = self.contigs & other.contigs
+    #     name = f"{self.name} & {other.name}"
+    #     origin = "intersection"
 
-        return Bin(contigs, origin, name)
+    #     return Bin(contigs, origin, name)
 
 
     def add_length(self, length: int) -> None:
@@ -131,7 +130,7 @@ class Bin:
         """
         other_contigs = (o.contigs for o in others)
         contigs = self.contigs.intersection(*other_contigs)
-        name = f"{self.name} & {' & '.join([other.name for other in others])}"
+        name = f"{self.id} & {' & '.join([str(other.id) for other in others])}"
         origin = "intersec"
 
         return Bin(contigs, origin, name)
@@ -145,7 +144,7 @@ class Bin:
         """
         other_contigs = (o.contigs for o in others)
         contigs = self.contigs.difference(*other_contigs)
-        name = f"{self.name} - {' - '.join([other.name for other in others])}"
+        name = f"{self.id} - {' - '.join([str(other.id) for other in others])}"
         origin = "diff"
 
         return Bin(contigs, origin, name)
@@ -159,7 +158,7 @@ class Bin:
         """
         other_contigs = (o.contigs for o in others)
         contigs = self.contigs.union(*other_contigs)
-        name = f"{self.name} | {' | '.join([other.name for other in others])}"
+        name = f"{self.id} | {' | '.join([str(other.id) for other in others])}"
         origin = "union"
 
         return Bin(contigs, origin, name)
@@ -234,7 +233,7 @@ def get_bins_from_directory(bin_dir: str, set_name: str, fasta_extensions: Set[s
 
 
 
-def parse_bin_directories(bin_name_to_bin_dir: Dict[str, str], fasta_extensions:Set[str]) -> Dict[str, list]:
+def parse_bin_directories(bin_name_to_bin_dir: Dict[str, str], fasta_extensions:Set[str]) -> Dict[str, Set[Bin]]:
     """
[…]
