+++ tests/bin_manager_test.py
@@ -8,6 +8,8 @@ import pytest
 from binette import bin_manager
 import networkx as nx
 
+import logging
+
 def test_get_all_possible_combinations():
     input_list = ["2", "3", "4"]
     expected_list = [("2", "3"), ("2", "4"), ("3", "4"), ("2", "3", "4")]
@@ -21,6 +23,7 @@ def example_bin_set1():
     bin2 = bin_manager.Bin(contigs={"3", "4"}, origin="test1", name="bin2")
     bin3 = bin_manager.Bin(contigs={"5"}, origin="test1", name="bin2")
     return {bin1, bin2, bin3}
+
 @pytest.fixture
 def example_bin_set2():
     bin1 = bin_manager.Bin(contigs={"1", "2", "3"}, origin="test2", name="binA")
@@ -96,13 +99,13 @@ def test_add_quality():
     
 
 
-def test_two_bin_intersection():
-    bin1 = bin_manager.Bin(contigs={"1", "2", "e", "987"}, origin="test1", name="bin1")
-    bin2 = bin_manager.Bin(contigs={"1", "e", "2", "33"}, origin="test2", name="binA")
+# def test_two_bin_intersection():
+#     bin1 = bin_manager.Bin(contigs={"1", "2", "e", "987"}, origin="test1", name="bin1")
+#     bin2 = bin_manager.Bin(contigs={"1", "e", "2", "33"}, origin="test2", name="binA")
 
-    bin_intersection = bin1 & bin2
+#     bin_intersection = bin1 & bin2
 
-    assert bin_intersection == bin_manager.Bin({"1", "2", "e"}, "", "")
+#     assert bin_intersection == bin_manager.Bin({"1", "2", "e"}, "", "")
 
 
 def test_multiple_bins_intersection():
@@ -159,7 +162,7 @@ def test_bin_union2():
     # Check the result
     expected_contigs = {'contig1', 'contig2', 'contig3', 'contig4', 'contig5'}
     expected_name = 'bin1 | bin2 | bin3'
-    expected_origin = 'union'
+    expected_origin = {'union'}
 
     assert union_bin.contigs == expected_contigs
     assert union_bin.name == expected_name
@@ -308,7 +311,7 @@ def test_get_contigs_in_bins():
 
     contigs = bin_manager.get_contigs_in_bins(bin_set)
 
-    assert contigs == {"c1", "c2", "c3", "c4", "c18"}
+    assert set(contigs) == {"c1", "c2", "c3", "c4", "c18"}
 
 
 def test_dereplicate_bin_sets():
@@ -467,12 +470,32 @@ def test_parse_contig2bin_tables(tmp_path):
     for name, expected in expected_bins.items():
         assert name in result_bin_dict
         assert len(result_bin_dict[name]) == len(expected)
-        for result_bin, expected_bin in zip(result_bin_dict[name], expected):
-            assert result_bin.contigs == expected_bin.contigs
-            assert result_bin.name == expected_bin.name
-            assert result_bin.origin == expected_bin.origin
+        for result_bin in result_bin_dict[name]:
+            assert result_bin in expected
 
 
+def test_parse_contig2bin_tables_with_duplicated_bins(tmp_path, caplog):
+    # Create temporary contig-to-bin tables for testing
+    test_tables = {
+        "set1": [
+            "# Sample contig-to-bin table for bin1",
+            "contig1\tbin1",
+            "contig2\tbin1",
+            "contig3\tbin2",
+            "contig3\tbin3",
+        ]
+    }
+
+    # Create temporary files for contig-to-bin tables
+    for name, content in test_tables.items():
+        table_path = tmp_path / f"test_{name}_contig2bin_table.txt"
+        table_path.write_text("\n".join(content))
+
+    # Call the function to parse contig-to-bin tables
[…]
