### 2024-08-28 bff2899
update test with new bin name
