### 2024-08-28 99fafaa
improve fct to infer bin set name
