+++ binette/io_manager.py
@@ -1,36 +1,97 @@
 import logging
-import os
 import pyfastx
-from typing import Iterable, List, Dict
+from typing import Iterable, List, Dict, Tuple
 import csv
 
 from binette.bin_manager import Bin
 
+from pathlib import Path
 
-def infer_bin_name_from_bin_inputs(input_bins: List[str]) -> Dict[str, str]:
+def get_paths_common_prefix_suffix(paths: List[Path]) -> Tuple[List[str], List[str], List[str]]:
     """
-    Infer bin names from a list of bin input directories.
+    Determine the common prefix parts, suffix parts, and common extensions of the last part of a list of pathlib.Path objects.
 
-    :param input_bins: List of input bin directories.
-    :return: Dictionary mapping inferred bin names to their corresponding directories.
+    :param paths: List of pathlib.Path objects.
+    :return: A tuple containing three lists:
+             - The common prefix parts.
+             - The common suffix parts.
+             - The common extensions of the last part of the paths.
     """
-    logging.debug("Inferring bin names from input bins:")
+    # Extract parts for all paths
+    parts = [list(path.parts) for path in paths]
+    
+    # Find the common prefix
+    if not parts:
+        return [], [], []
+    
+    # Initialize common prefix and suffix lists
+    common_prefix = list(parts[0])
+    common_suffix = list(parts[0])
+    # Determine common prefix
+    for part_tuple in parts[1:]:
+        common_prefix_length = min(len(common_prefix), len(part_tuple))
+        common_prefix = [common_prefix[i] for i in range(common_prefix_length) if common_prefix[:i+1] == part_tuple[:i+1]]
+        if not common_prefix:
+            break
 
-    commonprefix_len = len(os.path.commonprefix(input_bins))
-    reversed_strings = [s[::-1] for s in input_bins]
-    commonsufix_len = len(os.path.commonprefix(reversed_strings))
+    # Determine common suffix
+    for part_tuple in parts[1:]:
+        common_suffix_length = min(len(common_suffix), len(part_tuple))
+        common_suffix = [common_suffix[-i] for i in range(1, common_suffix_length + 1) if common_suffix[-i:] == part_tuple[-i:]]
+        if not common_suffix:
+            break
+    if len(parts) > 1:
+        common_suffix.reverse()
 
-    bin_name_to_bin_dir = {d[commonprefix_len: len(d) - commonsufix_len]: d for d in input_bins}
+    # Determine common extensions of the last part of the paths
+    if len(paths) == 1:
+        common_extensions = paths[0].suffixes
+    else:
+        common_extensions = list(paths[0].suffixes)
+        for path in paths[1:]:
+            common_extension_length = min(len(common_extensions), len(path.suffixes))
+            common_extensions = [common_extensions[i] for i in range(common_extension_length) if common_extensions[i] == path.suffixes[i]]
+            if not common_extensions:
+                break
+    
+    return common_prefix, common_suffix, common_extensions
+    
+def infer_bin_set_names_from_input_paths(input_bins: List[Path]) -> Dict[str, Path]:
+    """
+    Infer bin set names from a list of bin input directories or files.
 
-    logging.debug(f"Input bins:  {' '.join(input_bins)}")
-    logging.debug(f"Common prefix to remove: {os.path.commonprefix(reversed_strings)[::-1]}")
-    logging.debug(f"Common suffix to remove: {os.path.commonprefix(input_bins)}")
+    :param input_bins: List of input bin directories or files.
[…]
