+++ tests/io_manager_test.py
@@ -36,18 +36,130 @@ def test_infer_bin_name_from_bin_inputs():
     ]
 
     # Call the function
-    result = io_manager.infer_bin_name_from_bin_inputs(input_bins)
+    result = io_manager.infer_bin_set_names_from_input_paths(list(map(Path, input_bins)))
 
     # Define the expected output
     expected_result = {
-        '1': '/path/to/bin1',
-        '2': '/path/to/bin2',
-        '3': '/path/to/bin3'
+        'bin1': Path('/path/to/bin1'),
+        'bin2': Path('/path/to/bin2'),
+        'bin3': Path('/path/to/bin3')
     }
 
     # Check if the output matches the expected dictionary
     assert result == expected_result
 
+def test_infer_bin_name_from_single_path():
+    # Mock input data
+    input_bins = [
+        '/path/to/bin1',
+    ]
+
+    # Call the function
+    result = io_manager.infer_bin_set_names_from_input_paths(list(map(Path, input_bins)))
+
+    # Define the expected output
+    expected_result = {
+        '/path/to/bin1': Path('/path/to/bin1'),
+    }
+
+    # Check if the output matches the expected dictionary
+    assert result == expected_result
+
+    
+def test_infer_bin_name_from_bin_table_inputs():
+    # Mock input data
+    input_bins = [
+        '/path/to/bin1.tsv',
+        '/path/to/bin2.tsv',
+        '/path/to/bin3.tsv'
+    ]
+
+    # Call the function
+    result = io_manager.infer_bin_set_names_from_input_paths(list(map(Path, input_bins)))
+
+    # Define the expected output
+    expected_result = {
+        'bin1': Path('/path/to/bin1.tsv'),
+        'bin2': Path('/path/to/bin2.tsv'),
+        'bin3': Path('/path/to/bin3.tsv')
+    }
+
+    # Check if the output matches the expected dictionary
+    assert result == expected_result
+
+
+def test_infer_bin_name_from_bin_table_with_different_ext():
+    # Mock input data
+    input_bins = [
+        '/path/to/bin1.tsv',
+        '/path/to/bin2.tsv',
+        '/path/to/bin3.txt'
+    ]
+
+    # Call the function
+    result = io_manager.infer_bin_set_names_from_input_paths(list(map(Path, input_bins)))
+
+    # Define the expected output
+    expected_result = {
+        'bin1.tsv': Path('/path/to/bin1.tsv'),
+        'bin2.tsv': Path('/path/to/bin2.tsv'),
+        'bin3.txt': Path('/path/to/bin3.txt')
+    }
+
+    # Check if the output matches the expected dictionary
+    assert result == expected_result
+
+def test_infer_bin_name_from_bin_table_with_different_dir():
+    # Mock input data
+    input_bins = [
+        '/path/to/bins',
+        '/path2/result_bins',
+        '/path2/result/bins',
+    ]
+
+    # Call the function
+    result = io_manager.infer_bin_set_names_from_input_paths(list(map(Path, input_bins)))
+
+    # Define the expected output
+    expected_result = {
+        'path/to/bins' : Path('/path/to/bins'),
+        'path2/result_bins': Path('/path2/result_bins'),
+        'path2/result/bins': Path('/path2/result/bins'),
+    }
+
+    # Check if the output matches the expected dictionary
+    assert result == expected_result
+    
+def test_get_paths_common_prefix_suffix():
+    # Test case 1: No paths provided
+    assert io_manager.get_paths_common_prefix_suffix([]) == ([], [], [])
+
+    # # Test case 2: Single path
[…]
