### 2024-08-29 b7ee3f7
apply pathlib in tests
