+++ tests/main_binette_test.py
@@ -11,7 +11,7 @@ from unittest.mock import patch, MagicMock
 from collections import Counter
 from tests.bin_manager_test import create_temp_bin_directories, create_temp_bin_files
 from argparse import ArgumentParser
-
+from pathlib import Path
 
 @pytest.fixture
 def bins():
@@ -62,7 +62,7 @@ def test_select_bins_and_write_them(tmp_path, tmpdir, bins):
 
     # Run the function with test data
     selected_bins = select_bins_and_write_them(
-        set(bins), str(contigs_fasta), final_bin_report, min_completeness=60, index_to_contig=index_to_contig, outdir=str(outdir), debug=True
+        set(bins), contigs_fasta, Path(final_bin_report), min_completeness=60, index_to_contig=index_to_contig, outdir=outdir, debug=True
     )
 
     # Assertions to check the function output or file existence
@@ -104,11 +104,11 @@ def test_manage_protein_alignement_resume(tmp_path):
 
         # Run the function with test data
         contig_to_kegg_counter, contig_to_genes = manage_protein_alignement(
-            faa_file=str(faa_file),
-            contigs_fasta="contigs_fasta",
+            faa_file=Path(faa_file),
+            contigs_fasta=Path("contigs_fasta"),
             contig_to_length=contig_to_length,
-            contigs_in_bins={},
-            diamond_result_file="diamond_result_file",
+            contigs_in_bins=set(),
+            diamond_result_file=Path("diamond_result_file"),
             checkm2_db=None,
             threads=1,
             resume=True,
@@ -149,11 +149,11 @@ def test_manage_protein_alignement_not_resume(tmpdir, tmp_path):
         # Call the function
 
         contig_to_kegg_counter, contig_to_genes = manage_protein_alignement(
-            faa_file=str(faa_file),
-            contigs_fasta=contigs_fasta,
+            faa_file=Path(faa_file),
+            contigs_fasta=Path(contigs_fasta),
             contig_to_length=contig_to_length,
-            contigs_in_bins={},
-            diamond_result_file=diamond_result_file,
+            contigs_in_bins=set(),
+            diamond_result_file=Path(diamond_result_file),
             checkm2_db=None,
             threads=1,
             resume=True,
@@ -180,7 +180,7 @@ def test_parse_input_files_with_contig2bin_tables(tmp_path):
     fasta_file.write_text(fasta_file_content)
 
     # Call the function and capture the return values
-    bin_set_name_to_bins, original_bins, contigs_in_bins, contig_to_length = parse_input_files(None, [str(bin_set1), str(bin_set2)], str(fasta_file))
+    bin_set_name_to_bins, original_bins, contigs_in_bins, contig_to_length = parse_input_files(None, [bin_set1, bin_set2], fasta_file)
 
 
     # # Perform assertions on the returned values
@@ -190,7 +190,7 @@ def test_parse_input_files_with_contig2bin_tables(tmp_path):
     assert isinstance(contig_to_length, dict)
 
 
-    assert set(bin_set_name_to_bins) == {'1', "2"}
+    assert set(bin_set_name_to_bins) == {'bin_set1', "bin_set2"}
     assert len(original_bins) == 4
     assert contigs_in_bins == {"contig1","contig2", "contig3","contig4"}
     assert len(contig_to_length) == 4
@@ -206,12 +206,12 @@ def test_parse_input_files_with_contig2bin_tables_with_unknown_contig(tmp_path):
     fasta_file.write_text(fasta_file_content)
 
     with pytest.raises(ValueError):
[…]
