### 2024-08-29 f3abada
output original bin metric files
