+++ binette/main.py
@@ -17,7 +17,7 @@ import os
 import binette
 from binette import contig_manager, cds, diamond, bin_quality, bin_manager, io_manager as io
 from typing import List, Dict, Optional, Set, Tuple, Union, Sequence, Any
-
+from pathlib import Path
 
 def init_logging(verbose, debug):
     """Initialise logging."""
@@ -85,6 +85,7 @@ def parse_arguments(args):
         "-d",
         "--bin_dirs",
         nargs="+",
+        type=Path,
         action=UniqueStore,
         help="List of bin folders containing each bin in a fasta file.",
     )
@@ -94,11 +95,12 @@ def parse_arguments(args):
         "--contig2bin_tables",
         nargs="+",
         action=UniqueStore,
+        type=Path,
         help="List of contig2bin table with two columns separated\
             with a tabulation: contig, bin",
     )
 
-    input_group.add_argument("-c", "--contigs", required=True, help="Contigs in fasta format.")
+    input_group.add_argument("-c", "--contigs", required=True, type=Path, help="Contigs in fasta format.")
 
     # Other parameters category
     other_group = parser.add_argument_group('Other Arguments')
@@ -113,7 +115,7 @@ def parse_arguments(args):
 
     other_group.add_argument("-t", "--threads", default=1, type=int, help="Number of threads to use.")
 
-    other_group.add_argument("-o", "--outdir", default="results", help="Output directory.")
+    other_group.add_argument("-o", "--outdir", default=Path("results"), type=Path, help="Output directory.")
 
     other_group.add_argument(
         "-w",
@@ -135,8 +137,9 @@ def parse_arguments(args):
 
     other_group.add_argument(
         "--checkm2_db",
+        type=Path,
         help="Provide a path for the CheckM2 diamond database. "
-        "By default the database set via <checkm2 database> is used.",
+        "By default the database set via <checkm2 database> is used."
     )
 
     other_group.add_argument("--low_mem", help="Use low mem mode when running diamond", action="store_true")
@@ -157,9 +160,9 @@ def parse_arguments(args):
     args = parser.parse_args(args)
     return args
 
-def parse_input_files(bin_dirs: List[str], 
-                      contig2bin_tables: List[str],
-                      contigs_fasta: str,
+def parse_input_files(bin_dirs: List[Path], 
+                      contig2bin_tables: List[Path],
+                      contigs_fasta: Path,
                       fasta_extensions:Set[str] = {".fasta", ".fna", ".fa"}) -> Tuple[Dict[str, Set[bin_manager.Bin]], Set[bin_manager.Bin], Set[str], Dict[str, int]]:
     """
     Parses input files to retrieve information related to bins and contigs.
@@ -178,11 +181,11 @@ def parse_input_files(bin_dirs: List[str],
 
     if bin_dirs:
         logging.info("Parsing bin directories.")
-        bin_name_to_bin_dir = io.infer_bin_name_from_bin_inputs(bin_dirs)
+        bin_name_to_bin_dir = io.infer_bin_set_names_from_input_paths(bin_dirs)
         bin_set_name_to_bins = bin_manager.parse_bin_directories(bin_name_to_bin_dir, fasta_extensions)
     else:
         logging.info("Parsing bin2contig files.")
-        bin_name_to_bin_table = io.infer_bin_name_from_bin_inputs(contig2bin_tables)
+        bin_name_to_bin_table = io.infer_bin_set_names_from_input_paths(contig2bin_tables)
         bin_set_name_to_bins = bin_manager.parse_contig2bin_tables(bin_name_to_bin_table)
 
[…]
