### 2024-09-03 956cb17
add first draft of the tutorial
