+++ docs/tutorial/binette.md
@@ -1,0 +1,7 @@
+## Run Binette 
+
+```{code-cell} bash
+
+binette --bin_dirs maxbin2/ metabat2/ semibin2/output_bins/ concoct/bins/ -c Kickstart.megahit/R1.contigs.fa --verbose  -t 12 -o binette_results
+
+```
