+++ docs/tutorial/tutorial_main.md
@@ -1,0 +1,52 @@
+
+# Tutorial 
+
+The goal of this tutorial is to show an example of commands on how Binette can be used on real data. We will start ou journey from metagenomics reads that we  gonna download, then we will assemble these reads in contigs that we will bin with different binning tool. I finally we will use Binette to refine those bins.
+
+
+
+```{toctree}
+:caption: 'Tutorial'
+:maxdepth: 2
+
+set_env_and_get_data
+assembly
+binning
+binette
+analyse_binette_result.ipynb
+```
+
+
+<!-- 
+```{include} ./set_env_and_get_data.md
+```
+
+
+
+```{include} ./assembly.md
+```
+
+```{include} ./binning.md
+```
+
+```{include} ./binette.md
+```
+
+```{include} ./analyse_binette_result.ipynb
+``` -->
+
+
+<!-- ### Compare binette results with intial bins
+
+
+
+## Compare Binette and Das Tool 
+
+
+### Run Das Tool 
+
+
+### Compare binette results with intial bins
+
+
+ -->
