+++ pyproject.toml
@@ -44,8 +44,11 @@ doc = [
     "readthedocs-sphinx-search==0.3.1",
     "sphinx-autobuild==2021.3.14",
     "myst-parser==1.0.0",
-    "docutils==0.18.1"
-]
+    "docutils==0.18.1",
+    "myst-nb",
+    "nbsphinx"
+     ]
+     
 dev = [
     "pytest>=7.0.0",
     "pytest-cov"

[…]
