### 2024-09-03 a1f983c
adjust nbsphinx version in doc deps
