+++ pyproject.toml
@@ -45,7 +45,8 @@ doc = [
     "sphinx-autobuild==2021.3.14",
     "myst-parser==1.0.0",
     "docutils==0.18.1", #"myst-nb",
-    "nbsphinx==0.9.5"
+    "nbsphinx==0.9.5",
+    nbsphinx_link==1.3.0
      ]
 
 dev = [
