+++ docs/conf.py
@@ -55,7 +55,8 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'build', "api"]
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'sphinx_rtd_theme' #'alabaster' # 
+# html_theme = 'sphinx_rtd_theme' #'alabaster' # 
+html_theme = 'sphinx_book_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
