+++ pyproject.toml
@@ -46,7 +46,8 @@ doc = [
     "myst-parser==1.0.0",
     "docutils==0.18.1", #"myst-nb",
     "nbsphinx==0.9.5",
-    "nbsphinx_link==1.3.0"
+    "nbsphinx_link==1.3.0",
+    "sphinx-book-theme==1.0.1"
      ]
 
 dev = [
