### 2024-09-03 de93964
add mermaid deps
