+++ pyproject.toml
@@ -47,7 +47,8 @@ doc = [
     "docutils==0.18.1", #"myst-nb",
     "nbsphinx==0.9.5",
     "nbsphinx_link==1.3.0",
-    "sphinx-book-theme==1.0.1"
+    "sphinx-book-theme==1.0.1",
+    "sphinxcontrib.mermaid"
      ]
 
 dev = [
