+++ docs/conf.py
@@ -61,7 +61,7 @@ nbsphinx_execute = 'never'
 # `path/to/file:heading` instead of just `heading`
 autosectionlabel_prefix_document = True
 
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'build', "api"]
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'build', "api", "jupyter_execute"]
 
 
 
@@ -69,7 +69,7 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'build', "api"]
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 # html_theme = 'sphinx_rtd_theme' #'alabaster' # 
-html_theme = 'sphinx_book_theme'
+html_theme = 'sphinx_rtd_theme' #'sphinx_book_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
