+++ docs/tutorial/tutorial_main.md
@@ -1,59 +1,30 @@
 
 # Tutorial 
 
-In this tutorial, we'll walk through a practical example of how to use Binette with real data. We'll start by downloading metagenomics reads and then assemble these reads into contigs. Next, we'll use different binning tools to group the contigs. Finally, we'll use Binette to refine these bins and improve our results.
-
-```{mermaid}
----
-title: "Tutorial Overview:"
-align: center
----
-
-%%{init: {'theme':'default'}}%%
-
-graph LR
-
-  A[Download Metagenomics Reads] --> B
-  B[Assemble Reads into Contigs] --> c
-          subgraph Pangenome creation
-            a:::workflow
-            c:::workflow
-            g:::workflow
-            p:::workflow
-            a("annotate") --> c
-            c(cluster) --> g(graph)
-            g(graph) --> p(partition)
-        end
-
-
-  C[Bin Contigs with Binning Tools] --> D[Refine Bins with Binette]
-
-
-        
-    classDef panrgp fill:#4066d4
-    classDef panmodule fill:#d44066
-    classDef workflow fill:#d4ae40
-
-
-```
+In this tutorial, we'll walk through a practical example of how to use Binette with real data. We'll start by downloading metagenomics reads and then assemble these reads into contigs. Next, we'll use different binning tools to group the contigs into bins. Finally, we'll use Binette to refine these bins.
 
 ```{mermaid}
 
----
+--- 
 title: "Tutorial Overview:"
 align: center
+
+config:
+  look: handDrawn
+  theme: neutral
 ---
 
 
 graph TD
 
-    i[Get Metagenomics Reads] --> B[Assembly & Reads alignment]
+    i[metagenomics reads] --> B[assembly]
 
 
-    B --> metabat2 --> r[Binette]
+    B --> metabat2 --> r[binette]
     B --> maxbin2 --> r
     B --> concoct --> r
     B --> semibin2 --> r
+    r --> f[final bins]
     
         subgraph Binning
             metabat2:::binning
@@ -62,8 +33,8 @@ graph TD
             semibin2:::binning
         end
 
-        
-    classDef binning fill:#d4ae40
+            
+        classDef binning fill:#d4ae40
 
 
 ```
@@ -79,6 +50,7 @@ assembly
 binning
 binette
 analyse_binette_result.ipynb
+analyse_binette_result.myst
 ```
 
 
