+++ docs/api/modules.md
@@ -1,7 +1,0 @@
-# binette
-
-```{toctree}
-:maxdepth: 4
-
-binette
-```
