+++ docs/api/api_ref.md
@@ -3,6 +3,5 @@
 ```{toctree}
 :maxdepth: 2
 binette
-indice_and_table
 ```
 
