### 2024-09-04 716e2e5
update usage with input bin sets report new output
