+++ pyproject.toml
@@ -49,7 +49,7 @@ doc = [
     "nbsphinx_link==1.3.0",
     "sphinx-book-theme==1.0.1",
     "sphinxcontrib.mermaid",
-    "sphinx_togglebutton=0.3.2"
+    "sphinx_togglebutton==0.3.2"
      ]
 
 dev = [
