### 2024-09-04 62cd86d
add deps to env yml
