+++ docs/tutorial/binette.md
@@ -11,6 +11,14 @@ binette --bin_dirs maxbin2/ metabat2/ semibin2/output_bins/ concoct/bins/ \
         --verbose -t 12 -o binette_results
 ```
 
+```{admonition} ⌛ Expected Time
+:class: note
+:class: dropdown
+
+This process should talke around 9 minutes to complete.
+```
+
+
 Once Binette completes, the `binette_results` directory should have the following structure:
 
 ```
