### 2024-09-04 46cd347
fix typo in assembly tuto doc
