### 2024-09-04 91c3452
improve env file
