+++ binder/environment.yml
@@ -1,8 +1,8 @@
+name: binder_tutorial_env
 channels:
   - conda-forge
-  - defaults
 dependencies:
   - jupyter                # Jupyter notebook for interactive analysis
-  - pandas=1               # Data manipulation and analysis
+  - pandas               # Data manipulation and analysis
   - plotly=5               # Interactive graphing
-  - nbgitpuller
+  # - nbgitpuller
