+++ binder/requirements.txt
@@ -1,0 +1,3 @@
+jupyter                # Jupyter notebook for interactive analysis
+pandas               # Data manipulation and analysis
+plotly               # Interactive graphing
