### 2024-09-04 8aec082
remvoe binder env as it does not work
