### 2024-09-04 888923c
update tuto env file for binette 1.0.2
