### 2024-09-04 genotoul-bioinfo/Binette#19 @JeanMainguy
Hi,

The tutorial is now live! You can check it out here: https://binette.readthedocs.io/en/latest/tutorial/tutorial_main.html.

I’ve used the data you suggested from the ncezid-biome/datasets.

Also, I've released a new minor version that includes a new feature: it now generates quality reports for the input bins. Since the quality assessment is already part of the process, these new reports make it easier to compare the original input bins with the results produced by Binette.
