### 2024-09-05 ecb27b2
update readme with input bins report detail
