### 2024-09-06 44dd46d
give hint on checkm2 db install
