### 2024-09-06 #6782 @JeanMainguy
Hi,
I have added the tutorial to Binette documentation. You can check it out here: https://binette.readthedocs.io/en/latest/tutorial/tutorial_main.html

We plan to submit the metagWGS paper next week (hopefully) and will also upload the manuscript to BioRxiv at that time. I’ll keep you updated once the submission is complete and when I’ve updated the paper references accordingly.
