### 2024-09-19 51dccb0
fix checkm2 typo
