### 2024-09-19 #6782 @JeanMainguy
Hi @arfon ,
I just released a new version of Binette and published the archive on zenodo : https://doi.org/10.5281/zenodo.13789302
