### 2024-09-26 #6782 @JeanMainguy
Hi @arfon,

I forgot to mention that I’ve completed a final read-through of the paper before release and updated the metagWGS reference with its preprint DOI, as suggested earlier.

Is there anything else you need from my side to move this forward?

Thanks!
